requirements.txt
asyncssh==2.18.0
python-dotenv==1.0.1
tk
aiohttp==3.11.7
aiohttp-socks==0.9.1

#--- Chrome ---
selenium==4.27.1
webdriver-manager==4.0.2